Consolidate deeplink handling logic into useAddSite hook
